backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
Pillow==10.0.0
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
